requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
streamlit
requests